Use ActiveMQ instead of AMQ
