+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-client/.gitignore
