+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-client/src/main/java/org/acme/schooltimetabling/persistence/TimeTableRepository.java
