+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-client/src/main/resources/META-INF/resources/index.html
