+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-common/.gitignore
