+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-common/src/main/java/org/acme/schooltimetabling/domain/Lesson.java
