+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/.gitignore
