+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/main/docker/Dockerfile.native
@@ -1,0 +1,22 @@
+####
+# This Dockerfile is used in order to build a container that runs the Quarkus application in native (no JVM) mode
+#
+# Before building the docker image run:
+#
+# mvn package -Pnative -Dquarkus.native.container-build=true
+#
+# Then, build the image with:
+#
+# docker build -f src/main/docker/Dockerfile.native -t optaplanner/activemq-school-timetabling .
+#
+# Then run the container using:
+#
+# docker run -i --rm --network host optaplanner/activemq-school-timetabling
+#
+###
+FROM registry.access.redhat.com/ubi8/ubi-minimal
+WORKDIR /work/
+COPY target/*-runner /work/application
+RUN chmod 775 /work
+EXPOSE 8080
+CMD ["./application", "-Dquarkus.http.host=0.0.0.0"]
