+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/main/java/org/acme/schooltimetabling/solver/TimeTableConstraintProvider.java
