+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/main/resources/application.properties
@@ -1,0 +1,57 @@
+#
+# Copyright 2021 Red Hat, Inc. and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+########################
+# OptaPlanner properties
+########################
+
+# The solver runs for 10 seconds. To run for 5 minutes use "5m" and for 2 hours use "2h".
+quarkus.optaplanner.solver.termination.spent-limit=10s
+
+# To change how many solvers to run in parallel
+# optaplanner.solver-manager.parallel-solver-count=4
+# To run increase CPU cores usage per solver
+# quarkus.optaplanner.solver.move-thread-count=2
+
+# To detect common bugs in your code
+# quarkus.optaplanner.solver.environment-mode=FULL_ASSERT
+
+# XML file for power tweaking, defaults to solverConfig.xml (directly under src/main/resources)
+# quarkus.optaplanner.solver-config-xml=org/.../timeTableSolverConfig.xml
+
+########################
+# Test overrides
+########################
+
+# Effectively disable this termination in favor of the best-score-limit
+%test.quarkus.optaplanner.solver.termination.spent-limit=1h
+%test.quarkus.optaplanner.solver.termination.best-score-limit=0hard/*soft
+
+########################
+# Messaging
+########################
+
+amqp-username=quarkus
+amqp-password=quarkus
+
+# Configure the ActiveMQ source to read from it
+mp.messaging.incoming.solver_request.connector=smallrye-amqp
+mp.messaging.incoming.solver_request.durable=true
+mp.messaging.incoming.solver_request.failure-strategy=reject
+
+# Configure the ActiveMQ sink to write to it
+mp.messaging.outgoing.solver_response.connector=smallrye-amqp
+mp.messaging.outgoing.solver_response.durable=true
