+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/test/java/org/acme/schooltimetabling/TimeTableMessagingHandlerIT.java
