+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/test/resources/application.properties
@@ -1,0 +1,40 @@
+#
+# Copyright 2021 Red Hat, Inc. and/or its affiliates.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+########################
+# OptaPlanner properties
+########################
+
+%test.quarkus.optaplanner.solver.termination.spent-limit=1h
+%test.quarkus.optaplanner.solver.termination.best-score-limit=0hard/*soft
+
+########################
+# Messaging
+########################
+
+quarkus.artemis.url=tcp://localhost:61616
+
+# Configure the ActiveMQ source (we read from it)
+mp.messaging.incoming.solver_request.connector=smallrye-amqp
+mp.messaging.incoming.solver_request.durable=true
+mp.messaging.incoming.solver_request.failure-strategy=reject
+
+# Configure the ActiveMQ sink (we write to it)
+mp.messaging.outgoing.solver_response.connector=smallrye-amqp
+mp.messaging.outgoing.solver_response.durable=true
+
+# TODO: Remove after switching to fast jar by default.
+quarkus.package.type=fast-jar
