+++ activemq-quarkus-school-timetabling/activemq-quarkus-school-timetabling-solver/src/test/resources/broker.xml
