+++ activemq-quarkus-school-timetabling/docker-compose.yml
@@ -1,0 +1,18 @@
+# A docker compose file to start the ActiveMQ Artemis broker.
+# more details on https://github.com/vromero/activemq-artemis-docker.
+version: '2'
+
+services:
+
+  artemis:
+    image: vromero/activemq-artemis:2.15.0-alpine
+    ports:
+      - "8161:8161"
+      - "61616:61616"
+      - "5672:5672"
+    volumes:
+      - .:/var/lib/artemis/etc-override
+    environment:
+      ARTEMIS_USERNAME: quarkus
+      ARTEMIS_PASSWORD: quarkus
+
