--- amq-quarkus-school-timetabling/amq-quarkus-school-timetabling-client/src/test/java/org/acme/schooltimetabling/AMQTestResourceLifecycleManager.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2021 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.acme.schooltimetabling;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import io.quarkus.test.common.QuarkusTestResourceLifecycleManager;
-import io.smallrye.reactive.messaging.connectors.InMemoryConnector;
-
-public class AMQTestResourceLifecycleManager implements QuarkusTestResourceLifecycleManager {
-
-    @Override
-    public Map<String, String> start() {
-        Map<String, String> env = new HashMap<>();
-        Map<String, String> solverRequestSinkProperties = InMemoryConnector.switchOutgoingChannelsToInMemory("solver_request");
-        Map<String, String> solverResponseSourceProperties =
-                InMemoryConnector.switchIncomingChannelsToInMemory("solver_response");
-        env.putAll(solverRequestSinkProperties);
-        env.putAll(solverResponseSourceProperties);
-        return env;
-    }
-
-    @Override
-    public void stop() {
-        InMemoryConnector.clear();
-    }
-}
